+++ tests/testthat/test-utils.R
@@ -4,5 +4,5 @@ test_that("quantiles", {
   quantiles <- quantiles(1:10, 0.95)
   expect_identical(names(quantiles), c("estimate", "lower", "upper"))
   expect_equal(quantiles, c(5.5, 1.225, 9.775), check.attributes = FALSE)
-  expect_equal(instantaneous_to_interval(interval_to_instantaneous(0.5)), 0.5)
+  expect_equal(ypr_instant2interval(ypr_interval2instant(0.5)), 0.5)
 })