+++ NEWS.md
@@ -6,7 +6,8 @@
 - Increased upper limit for `Rk` from 10 to 100.
 - Added scaling parameter `R0`.
 - Renamed `a` to `Wa` and `b` to `Wb`.
-- Switched `nu` for `S` and added `Mb` to scale mortality by length.
+- Added `ypr_instant2interval()` and `ypr_interval2instant()` to convert mortality rates.
+- Switched `nu` for `M` and added `Mb` to scale mortality by length.
 - Added `ypr_ggplot2()` to do customisable bivariate plot.
 
 # ypr 0.0.1